Write transaction hook to calculate total number of shares a user has
- write a helper method and beforeCreate hook in the Transaction model
- In order to prevent having to query and filter the entire database
of transactions to reset a user's total holdings of stock in a given
company each time a transaction takes place, the hook locates the last
entry where the user sold or purchased that company's stock. It then adds
the number of shares being transacted in the current transaction and
updates that entry's total figure.